Don't call `ScenManager.getGroup` during `Scene.load`
Prior to changes to use the scene manager for adding objects to the scene, `FollowPlayer` called `Scene.getGroup` to get a reference to the player object. This doesn't work when using the scene manager, because groups are not immediately updated.

Changes introduced no longer call `Scenemanager.getGroup` inside of the constructor of `FollowPlayer`, which is called during `Scene.load`. Instead, a reference to the player is directly assigned to a public attribute of `FollowPlayer` in `MainScene`.